refactor(frontend): extract UsageInfo to shared location, useUsage accepts eventBus param
