--- apps/frontend/src/pages/chat/components/InfoBar/components/UsageInfo/index.ts
@@ -1,1 +1,0 @@
-export {UsageInfo} from './UsageInfo.js';
