+++ apps/frontend/src/pages/chat/components/UsageInfo/helpers/format-token-count.ts
