+++ apps/frontend/src/pages/chat/components/UsageInfo/index.ts
@@ -1,0 +1,2 @@
+export {useUsage} from './hooks/useUsage.js';
+export {UsageInfo} from './UsageInfo.js';
